requirements.txt
numpy
modal
requests
openai
python-dotenv
fastapi
uvicorn